Revert "Enable SSE at @configu/proxy (#697)"
This reverts commit b7551293c8f6a2c76c76ccda0987c16472bbe779.
