+++ packages/proxy/package.json
@@ -43,7 +43,6 @@
     "@configu/sdk": "workspace:*",
     "@fastify/bearer-auth": "^9.4.0",
     "@fastify/cors": "^9.0.1",
-    "@fastify/formbody": "^7.4.0",
     "@fastify/helmet": "^11.1.1",
     "@fastify/swagger": "^8.14.0",
     "@fastify/swagger-ui": "^4.0.0",
@@ -52,12 +51,9 @@
     "cosmiconfig": "^9.0.0",
     "fastify": "^4.28.1",
     "fastify-plugin": "^4.5.1",
-    "lodash": "^4.17.21",
-    "node-cron": "^3.0.3"
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@types/node": "22.10.6",
-    "@types/node-cron": "^3.0.11",
     "fastify-cli": "^6.2.1"
   }
 }
