--- packages/proxy/src/__tests__/.configu
@@ -1,10 +1,0 @@
-{
-  "stores": {
-    "testStore": {
-      "type": "json-file",
-      "configuration": {
-        "path": "./packages/proxy/src/__tests__/configs.json"
-      }
-    }
-  }
-}
