--- packages/proxy/src/__tests__/configs.json
@@ -1,12 +1,0 @@
-[
-  {
-    "key": "GREETING",
-    "set": "",
-    "value": "hey"
-  },
-  {
-    "key": "SUBJECT",
-    "set": "",
-    "value": "Change2"
-  }
-]
