--- packages/proxy/src/__tests__/schema.json
@@ -1,14 +1,0 @@
-{
-  "keys": {
-    "GREETING": {
-      "pattern": "^(hello|hey|welcome|hola|salute|bonjour|shalom)$",
-      "default": "hello"
-    },
-    "SUBJECT": {
-      "default": "world"
-    },
-    "MESSAGE": {
-      "description": "Generates a full greeting message"
-    }
-  }
-}
