--- packages/proxy/src/__tests__/start.cfgu.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://files.configu.com/schema/.cfgu.json",
-  "keys": {
-    "GREETING": {
-      "pattern": "^(hello|hey|welcome|hola|salute|bonjour|shalom)$",
-      "default": "hello"
-    },
-    "SUBJECT": {
-      "test": "expect($.value).to.be.a('string')",
-      "default": "world"
-    },
-    "MESSAGE": {
-      "const": "$.configs.GREETING.value, $.configs.SUBJECT.value!",
-      "description": "Generates a full greeting message"
-    }
-  }
-}
