+++ packages/proxy/src/routes.ts
@@ -1,5 +1,4 @@
 import { FastifyPluginAsync } from 'fastify';
-import cron from 'node-cron';
 import {
   _,
   ConfigSchema,
@@ -12,6 +11,7 @@ import {
   FromSchema,
 } from '@configu/sdk';
 import { ConfiguInterface } from '@configu/common';
+import { config } from './config';
 
 const body = {
   type: 'array',
@@ -21,12 +21,19 @@ const body = {
     required: ['store', 'set', 'schema'],
     additionalProperties: false,
     properties: {
-      store: { type: 'string', minLength: 1 },
-      set: { type: 'string' },
+      store: {
+        type: 'string',
+        minLength: 1,
+      },
+      set: {
+        type: 'string',
+      },
       schema: {
         type: 'object',
         properties: {
-          keys: { type: 'object' },
+          keys: {
+            type: 'object',
+          },
         },
       },
       configs: {
@@ -44,74 +51,8 @@ const ok = {
   additionalProperties: true,
 } as const;
 
-// Optional querystring interface with cron
-interface ExportQuerystring {
-  cron?: string;
-}
-
-export async function runExportAndGetResult(requestBody: FromSchema<typeof body>) {
-  const evalResToExport = await requestBody.reduce<Promise<EvalCommandOutput>>(
-    async (previousResult, { store, set, schema: { keys }, configs }) => {
-      const pipe = await previousResult;
-
-      const storeInstance = await ConfiguInterface.getStoreInstance(store);
-      if (!storeInstance) {
-        throw new Error(`store "${store}" not found`);
-      }
-
-      const setInstance = new ConfigSet(set);
-      // todo: fix this any
-      const schemaInstance = new ConfigSchema(keys as any);
-
-      const evalCmd = new EvalCommand({
-        store: storeInstance,
-        set: setInstance,
-        schema: schemaInstance,
-        configs,
-        pipe,
-      });
-
-      const { result } = await evalCmd.run();
-
-      await ConfiguInterface.backupEvalOutput({
-        storeName: store,
-        set: setInstance,
-        schema: schemaInstance,
-        evalOutput: result,
-      });
-
-      // // TODO: move backup logic to common
-      // const backupStoreInstance = await configuFile.getBackupStoreInstance(store);
-      // if (backupStoreInstance) {
-      //   const backupConfigs = _(evalRes.result)
-      //     .pickBy((entry) => entry.origin === EvaluatedConfigOrigin.Store)
-      //     .mapValues((entry) => entry.value)
-      //     .value();
-      //   await new UpsertCommand({
-      //     store: backupStoreInstance,
-      //     set: setInstance,
-      //     schema: schemaInstance,
-      //     configs: backupConfigs,
-      //   }).run();
-      // }
-
-      return result;
-    },
-    Promise.resolve({}),
-  );
-
-  const exportCmd = new ExportCommand({ pipe: evalResToExport });
-  const exportRes = await exportCmd.run();
-
-  return exportRes;
-}
-
 export const routes: FastifyPluginAsync = async (server, opts): Promise<void> => {
-  server.post<{
-    Body: FromSchema<typeof body>;
-    Reply: FromSchema<typeof ok>;
-    Querystring: ExportQuerystring;
-  }>(
+  server.post<{ Body: FromSchema<typeof body>; Reply: FromSchema<typeof ok> }>(
     '/export',
     {
       schema: {
@@ -125,79 +66,58 @@ export const routes: FastifyPluginAsync = async (server, opts): Promise<void> =>
       // // TODO: get the ConfiguFile instance from a shared location
       // const configuFile = await ConfiguFile.load(config.CONFIGU_CONFIG_FILE);
 
-      // If there's no cron in the query, do the one-time export and return JSON
-      if (!request.query.cron) {
-        const exportRes = await runExportAndGetResult(request.body);
+      const evalResToExport = await request.body.reduce<Promise<EvalCommandOutput>>(
+        async (previousResult, { store, set, schema: { keys }, configs }) => {
+          const pipe = await previousResult;
 
-        // TODO: consider if this is the right way to parse the result
-        // const parsedExportRes = JSON.parse(exportRes.result);
+          const storeInstance = await ConfiguInterface.getStoreInstance(store);
+          if (!storeInstance) {
+            throw new Error(`store "${store}" not found`);
+          }
+          const setInstance = new ConfigSet(set);
+          // todo: fix this any
+          const schemaInstance = new ConfigSchema(keys as any);
 
-        // Return one-time result
-        return exportRes;
-      }
+          const evalCmd = new EvalCommand({
+            store: storeInstance,
+            set: setInstance,
+            schema: schemaInstance,
+            configs,
+            pipe,
+          });
+          const { result } = await evalCmd.run();
+          await ConfiguInterface.backupEvalOutput({
+            storeName: store,
+            set: setInstance,
+            schema: schemaInstance,
+            evalOutput: result,
+          });
 
-      // Otherwise, we are in SSE mode, triggered by the presence of a "cron" query parameter
-      if (!cron.validate(request.query.cron)) {
-        throw new Error(`Invalid cron expression: "${request.query.cron}"`);
-      }
+          // // TODO: move backup logic to common
+          // const backupStoreInstance = await configuFile.getBackupStoreInstance(store);
+          // if (backupStoreInstance) {
+          //   const backupConfigs = _(evalRes.result)
+          //     .pickBy((entry) => entry.origin === EvaluatedConfigOrigin.Store)
+          //     .mapValues((entry) => entry.value)
+          //     .value();
+          //   await new UpsertCommand({
+          //     store: backupStoreInstance,
+          //     set: setInstance,
+          //     schema: schemaInstance,
+          //     configs: backupConfigs,
+          //   }).run();
+          // }
 
-      reply.raw.writeHead(200, {
-        'Content-Type': 'text/event-stream',
-        'Cache-Control': 'no-cache',
-        Connection: 'keep-alive',
-      });
+          return result;
+        },
+        Promise.resolve({}),
+      );
 
-      let eventId = 0;
-      function getNextId() {
-        if (eventId === Number.MAX_SAFE_INTEGER) {
-          eventId = 0; // Reset
-        }
-        eventId += 1;
-        return eventId;
-      }
-
-      // Helper to run existing logic and write to SSE
-      const runExportAndWriteSSE = async () => {
-        try {
-          const exportRes = await runExportAndGetResult(request.body);
-          const replyObj = {
-            id: getNextId(),
-            event: 'export',
-            data: exportRes,
-          };
-
-          reply.raw.write(`data: ${JSON.stringify(replyObj)}\n\n`);
-        } catch (error: any) {
-          const errMsg = error?.message || 'Unknown error';
-          reply.raw.write(`event: error\n`);
-          reply.raw.write(`data: ${errMsg}\n\n`);
-        }
-      };
-
-      const task = cron.schedule(request.query.cron, async () => {
-        await runExportAndWriteSSE();
-      });
-
-      // Run export once immediately before the cron schedule ticks
-      await runExportAndWriteSSE();
-
-      // Clean up if the client disconnects
-      request.raw.on('close', () => {
-        reply.raw.end();
-        task.stop();
-      });
-      request.raw.on('end', () => {
-        reply.raw.end();
-        task.stop();
-      });
-      request.raw.on('aborted', () => {
-        reply.raw.end();
-        task.stop();
-      });
-
-      // Because we’re streaming, keep the connection open (until client disconnects or an error occurs)
-      // and return the reply
-      return reply;
+      const exportCmd = new ExportCommand({ pipe: evalResToExport });
+      const exportRes = await exportCmd.run();
+      // TODO: consider if this is the right way to parse the result
+      // const parsedExportRes = JSON.parse(exportRes.result);
+      return reply.send(exportRes);
     },
   );
 };
